refactor helpers to separate file to be reusable
